Added Constant.dart
