+++ android/app/src/main/AndroidManifest.xml
@@ -1,7 +1,7 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.example.calcflutter.calc_flutter_d">
+    package="com.example.calcflutter.calc_flutter">
    <application
-        android:label="calc_flutter_d"
+        android:label="calc_flutter"
         android:icon="@mipmap/ic_launcher">
         <activity
             android:name=".MainActivity"
