+++ android/app/src/main/kotlin/com/example/calcflutter/calc_flutter_d/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.example.calcflutter.calc_flutter_d
+package com.example.calcflutter.calc_flutter
 
 import io.flutter.embedding.android.FlutterActivity
