+++ lib/util/Constant.dart
@@ -1,0 +1,22 @@
+class Constant {
+  static const String APP_NAME = "Flutter Calculator";
+
+  static const String BTN_CLEAR = "C";
+  static const String BTN_BACKSPACE = "⌫";
+  static const String BTN_MULTIPLY = "×";
+  static const String BTN_DIVIDE = "÷";
+  static const String BTN_DECIMAL = ".";
+  static const String BTN_MINUS = "-";
+  static const String BTN_PLUS = "+";
+  static const String BTN_EQUAL = "=";
+  static const String BTN_0 = "0";
+  static const String BTN_1 = "1";
+  static const String BTN_2 = "2";
+  static const String BTN_3 = "3";
+  static const String BTN_4 = "4";
+  static const String BTN_5 = "5";
+  static const String BTN_6 = "6";
+  static const String BTN_7 = "7";
+  static const String BTN_8 = "8";
+  static const String BTN_9 = "9";
+}
