+++ pubspec.yaml
@@ -1,4 +1,4 @@
-name: calc_flutter_d
+name: calc_flutter
 description: Flutter calculator
 
 # The following line prevents the package from being accidentally published to
